+++ screens/Ratings/RatingsScreen.tsx
@@ -1,39 +1,48 @@
 import { useState, useEffect } from "react";
 import { StyleSheet } from "react-native";
-import { getFirestore, collection, getDocs } from "firebase/firestore";
+import { getFirestore, collection, getDocs, query } from "firebase/firestore";
 
 import { Text, View } from "../../components/Themed";
 import { Categories } from "./Categories";
-import { db } from "../../App"
+import { db } from "../../firebase/firebaseConfig";
 
 export type Category = {
   title: String;
   quantity: number;
 };
 
-const categories: Category[] = [
-  {
-    title: "exercise",
-    quantity: 5,
-  },
-  {
-    title: "nofap",
-    quantity: 4,
-  },
-];
+// const categories: Category[] = [
+//   {
+//     title: "exercise",
+//     quantity: 5,
+//   },
+//   {
+//     title: "nofap",
+//     quantity: 4,
+//   },
+// ];
 
 async function getCategories(db: any) {
-  return await getDocs(collection(db, "category"));
-  // categorySnapshot.forEach((doc) => console.log(`${doc.id} => ${doc.data()}`));
+  const q = query(collection(db, "category"));
+  const snapshot = await getDocs(q);
+  return snapshot.docs.map((doc) => {
+    return doc.data();
+  });
 }
 
 export default function RatingsScreen() {
   const [selection, setSelection] = useState(null);
-  const [categories, setCategories] = useState([])
+  const [categories, setCategories] = useState([]);
+
+  getCategories(db).then((data) => {
+    setCategories(data);
+  });
+
+  console.log(categories)
 
   return (
     <View style={styles.container}>
-      {categories.length > 0 ? (
+      {categories.length ? (
         <Categories
           categories={categories}
           selection={selection}
